Moved data translator to data utils.
